+++ image-lint/bin/image-lint.js
@@ -7,13 +7,14 @@ import type {ExtendedOptions} from '../lib/args-helper.js';
 */
 
 const minimist = require('minimist'),
+	  chalk = require('chalk'),
 	  ArgsHelper = require('../lib/args-helper').default,
 	  linter = require('../lib/linter'),
 	  defaults = require('../lib/defaults').default,
 	  Linter = linter.default,
 	  DEFAULT_DIRECTORY = ['./'];
 
-var argument_config/*: minimistOptions */ = {
+const argument_config/*: minimistOptions */ = {
 	'boolean': [
 		'mismatch',
 		'duplicate',
@@ -42,15 +43,20 @@ var argument_config/*: minimistOptions */ = {
 	}
 };
 
-var argv = minimist(process.argv.slice(2), argument_config);
+const argv = minimist(process.argv.slice(2), argument_config);
+
+const EXIT_CODE_OK = 0;
+const EXIT_CODE_LINT_ERROR = 1;
 
 if (ArgsHelper.argv(argument_config, argv)) {
-	process.exitCode = 0;
+	process.exitCode = EXIT_CODE_OK;
 } else {
 	const MultiFinder = require('../lib/finder/multi');
 	const ImageIdentifier = require('../lib/ident');
 
 	let folder = argv._;
+	let error_count = 0;
+	let warning_count = 0;
 	const finder = new MultiFinder(ImageIdentifier.get_all_extensions(), ImageIdentifier.get_all_mimes());
 	const cli_linter = new Linter(finder);
 
@@ -62,8 +68,22 @@ if (ArgsHelper.argv(argument_config, argv)) {
 
 	cli_linter.lint(folder, argv)
 		.on('file.completed', (logger) => {
+			error_count += logger.get_error_count();
+			warning_count += logger.get_warning_count();
+
 			if (logger.is_printable()) {
 				console.log(logger.toString());
 			}
+		})
+		.on('linter.completed', () => {
+			const has_errors = (error_count > 0 || warning_count > 0);
+			const warnings = warning_count > 0 ? chalk.yellow(warning_count) : warning_count;
+			const errors = error_count > 0 ? chalk.red(error_count) : error_count;
+
+			if (has_errors) {
+				console.log(`\nTotal Warnings: ${ warnings }, Errors: ${ errors }`);
+			}
+
+			process.exitCode = has_errors ? EXIT_CODE_LINT_ERROR : EXIT_CODE_OK;
 		});
 }